Fix NPE in repl
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/sip-servlets@8641 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ sip-servlets-tomcat-jboss4/src/main/java/org/jboss/web/tomcat/service/session/JBossCacheSipManager.java
@@ -2076,7 +2076,9 @@ public class JBossCacheSipManager extends JBossCacheManager implements
 			}
 		} else {
 			//remove the dialog from the local stacks 
-			session.getSessionCreatingDialog().delete();
+			if(session.getSessionCreatingDialog() != null) {
+				session.getSessionCreatingDialog().delete();
+			}
 			
 			// Expire the session
 			// DON'T SYNCHRONIZE ON SESSION HERE -- isValid() and
